lightfm_service/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.11.4
lightfm==1.17
scikit-learn==1.3.2
